Reorganize credit card related components
