+++ src/components/CreditCardDetails/components/CardNumberInput/constants/index.js
@@ -1,4 +1,3 @@
 import VALIDATION_REGEXES from './validation-regexes';
 
-export * from './card-schemes';
 export { VALIDATION_REGEXES };
