+++ src/components/CreditCardDetails/components/CardNumberInput/index.js
@@ -1,6 +1,3 @@
 import CardNumberInput from './CardNumberInput';
 
-export * from './components';
-export { SCHEMES } from './constants';
-
 export default CardNumberInput;
