+++ src/components/CreditCardDetails/components/index.js
@@ -1,0 +1,4 @@
+import CardNumberInput from './CardNumberInput';
+import cardSchemeIcons from './scheme-icons';
+
+export { CardNumberInput, cardSchemeIcons };
