+++ src/components/CreditCardDetails/constants/index.js
@@ -1,0 +1,3 @@
+import * as schemes from './card-schemes';
+
+export { schemes };
